Web/JWT Kiddo/chall/challenge/requirements.txt
Flask
pyjwt[crypto]